ardop: Ignore ARQ frames when not connected
According to the ARDOP host specification, ARQ frames should only
contain "data received from a connected (ARQ) session and should be
error free". This is not the case, as both ARDOP_win and ARDOPc is
sending monitor logs as ARQ frames when the TNC is not idle.

This patch tries to ignore these frames when the tnc is idle. It may
not work every time, because of the asynchronous nature of the two
tcp sockets; One socket sending the connected/disconnected state updates
while the other one is sending ARQ frames. It seems like a race may
occure, but this is how I've been told to handle the issue.

I should also add that I have not been able to get a comment from Rick
on this matter, dispite multiple attempts to contact him on the ARDOP
developers mailing list.

Ref la5nta/pat#139