Avoid race when closing the ardop TNCs
The real fix would be to get rid of the race by using a mutex or some
other safe mechanism to guard against concurrent calls to Close().

Since I don't have time to test this properly, this may act as a quick
fix until there is time for the proper fix.

Closes #52