fbb: New attachment filename max length is 255
While implementing support for Winlink Forms we noticed that some forms
requires filenames > 50 characters, which violated the B2F
specification. This lead to a discussion with the Winlink Development
Team where they concluded that the hard limit should be increased to 255
characters.

This change should be backwards compatible with CMS and Winlink Express,
as neither of them did check the filename length against this 50
character limit.

Ref https://github.com/la5nta/pat/issues/135