fbb: Support different passwords for aux addresses
Previously we assumed that all auxiliary addresses was configured with
the same password as the user's account. Now we call
SecureLoginHandleFunc for each aux address with the address in
question as a function parameter. That will enable applications to
provide different passwords for each address.

This is a breaking change, since the SecureLoginHandleFunc previously
took no arguments.

Ref issue la5nta/pat#143