Add capability to send commands that return more than 1 result
The interface to rigctld before this commit could only handle commands that return either one or no results. The problem is that commands like "\get_mode" return two lines of output but don't indicate when they are finished. So we have to tell it how many lines to read; e.g., send get_mode and expect two lines of output. The outputs are packed into a string array, so for commands returning single outputs (like get_freq) the output is the result[0].

The exported functions are compatible with the libhamlib interface, even if the unexported internals are slightly different.
Remove all debug tracing

Hopefully, I'll be able to merge this back in later if necessary.