Add `bw` query param parsing to ARDOP
This will help implement https://github.com/la5nta/pat/issues/120. This generally wants the `bw` query parameter to match the string sent to the ARDOP TNC, e.g. `500MAX` or `2000FORCED`, but relaxes the API a bit by assuming `MAX` if no word is sent with the bandwidth value.