Throttle Write rate based on 'Y' frame
This only works reliably for Direwolf >= 1.5, but the 'y' frame
implementation in Direwolf is error prune and not reliable at all.

Need to confirm that this works with other modems, because Direwolf does
not follow the AGWPE spec for either frame types. We might need a
direwolf flag in config for this to work reliably with other modems on
inbound connections.